[fidl][examples] Adding echo-realm target.
The build file in //examples/fidl/echo-realm does not have
a target that matches the directory name. This causes confusion
when trying to refer to it in a build.

This adds a group named "echo-realm" which depends on the "echo_realm"
target.

Fixes: b/297069811
Change-Id: I94af1cf9821b5a36ba2601808bfbdbc94f517e03
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/906818
Fuchsia-Auto-Submit: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Ian McKellar <ianloic@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>