[time] Clean up canonical targets
Makes the target //src/sys/timekeeper:timekeeper a
canonical testonly target. It used to refer to the
timekeeper package.

This is no longer the case, and is no longer necessary.
Which is why it is being retracted now.

Bug: b/300289294
Change-Id: I6a266d7483c8a1c35c85a96e48ba736e95056c65
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/918649
Commit-Queue: Filip Filmar <fmil@google.com>
Reviewed-by: Mukesh Agrawal <quiche@google.com>
Fuchsia-Auto-Submit: Filip Filmar <fmil@google.com>