[netstack3] Implement ICMP sockets in terms of datagram module
In this change, `connect`, `bind` and `send_conn` are implemented for
the ICMP sockets in Core. The tests are testing the same behavior they
used to, except for sending or receiving, the old behavior is that the
buffer only represents the body of the ICMP echo, the new behavior is
that the buffer is the entire ICMP echo packet, including the ICMP
header.

Bug: 125482

Change-Id: Ie91490abb28393fc4c20ee59ca6630612c9e630a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/917534
Reviewed-by: Jeff Martin <martinjeffrey@google.com>
Commit-Queue: Zeling Feng <zeling@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Fuchsia-Auto-Submit: Zeling Feng <zeling@google.com>