[starnix] ptrace attach support
Supports PTRACE_ATTACH, PTRACE_KILL, and PTRACE_TRACEME.

This does not let you do anything particularly interesting with the
tracee.

This CL contains several changes to signal processing:

Previously, a thread group could be in a stop state, and
individual tasks could not.  Tasks would stop if their thread
group was in a stop state.

Now, thread groups and tasks both have stop states.  Calls to
waitpid() wait until a change to either the thread group's state
(e.g., which happens if a SIGSTOP or SIGCONT is sent) or an
individual task'state (which happens if it is traced, and, for
example, a signal gets delivered, which stops traced processes).

Stop states are further broken down by type, per ptrace(2). This CL
only supports signal-delivery-stop.

Bug: b/299348370

Change-Id: Ie57c3d2a5c23287f643d7f2e98d480fd9c686da5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/902166
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Jeremy Manson <jeremymanson@google.com>