[fidl] Include external flexible method result structs
Flexible two-way methods that don't have an error type specified are
actually implemented on the wire as a struct with a union that has a
success and a transport error member.

For composed protocols the definition of that top-level response struct
must be included in the external_struct_declarations IR field. It was
for methods with errors, but not for flexible methods.

Test: Added test-cases to the fidlc testdata, verified that they broke,
then verified that they're fixed.

Change-Id: Ic0fad1b51d56e11cb0e4209ba8c59541185e9b39
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/915439
Reviewed-by: Mitchell Kember <mkember@google.com>
Fuchsia-Auto-Submit: Ian McKellar <ianloic@google.com>
Commit-Queue: Ian McKellar <ianloic@google.com>