[fuchsia-controller] Fix result type import bug.
If there's a FIDL library that is being imported
and there's a struct that contains a channel to a composed protocol, and
the composed protocol has a method that returns a result,
fuchsia-controller would throw an exception. This fixes that and
includes a test reproducing this issue in test fidl code.

Bug: b/301178571

Change-Id: I75a54c035933c16915ca83ebaf941a4bbd2fe886
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/919196
Reviewed-by: Kevin Cho <chok@google.com>
Commit-Queue: Andrew Davies <awdavies@google.com>