[fxtest] Use separate script for matching
The matching we do is expensive, and the version of jellyfish we use for
approximate matching implements the algorithms in Python itself.

This change instead depends on the new dldist program for approximate
matching, each string is passed on a separate line and the matching
output lines are the distance measurement for that input line as a
decimal integer.

This change does the following:
- Call dldist for matching decisions.
- Calls to dldist are added as children of "Matching" events, matching
  is now asynchronous.
- An option is added to the event group payloads to indicate that status
  updates should skip outputting the (verbose list of) children for each
  matching operation.
- Console output is changed to respect the hide children option.
- dldist has an option to match the prefix logic that was previously
  implemented in Python.

Fixed: b/300157465
Change-Id: I46d6385d94b8233fd692d9d01cb2075ea20dfae7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/918217
Fuchsia-Auto-Submit: Christopher Johnson <crjohns@google.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>