[fidl] Put @transitional behind an allowlist
The @transitional attribute was added many years ago to simplify making
FIDL changes in the layer-cake era. We now have a more authentic
monorepo and @available to support out-of-tree changes.

It's easy to build incorrect software when using @transitional and we've
seen @transitional left on methods for more than four years. It's time
to transition to our shiny new glorious versioning world.

This adds an allow-list for the use of @transitional so that as we
eliminate legacy uses of it no more are added. Right now that's behind
the experimental flag transitional_allow_list because there are FIDL
files out of tree that use @transitional, but once those are gone this
should become default behavior.

Change-Id: I8196caac13269571f1b89ac5769fec0f133711cb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/919714
Commit-Queue: Ian McKellar <ianloic@google.com>
Reviewed-by: Mitchell Kember <mkember@google.com>