[ffx][overnet]Adding ABI check to overnet
This updates the runner script on the target to check a provided ABI
revision to the one used to compile the runner.

This result is then returned in a JSON object before starting the
Overnet communication.

Support is also included to fall back to invoking the legacy
runner if the abi revision flag is not known.

The result of the comparison is stored in the target object
so it can be used to inform the user of the compatibility status
and protocol versioning decisions can be made.

Bug: b/295899889

Change-Id: Idcd4b003fcb3ec89fddfc12b9122e9af33718663
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/901721
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>
API-Review: Miguel Flores <miguelfrde@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Casey Dahlin <sadmac@google.com>