[kernel][pager] Fix InvalidPagerSupplyPages on Vim3
The `InvalidPagerSupplyPages` test was failing on Vim3 because it tried
to create a physical VMO at address 0x0, which on a Vim3 is not within
the range of peripheral memory. This CL fixes the issue by using
`vmo_test::GetTestPhysVmo` to create the physical VMO in a valid region.

Change-Id: I354844654d4db587535b9ee0048c5884996214df
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/921539
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Anirudh Mathukumilli <rudymathu@google.com>