[starnix] Implement MADV_DONTFORK and MADV_WIPEONFORK for madvise()
This implements the DONTFORK and WIPEONFORK advice values for the
madvise() system call. DONTFORK marks a mapping such that it will not
create any mapping in a cloned address space. It can be undone by the
DOFORK advice. This advice follows a memory mapping through remapping
and extension via remap(). WIPEONFORK is a similar annotation that
replaces a private, anonymous mapping with a fresh mapping of zeroed
memory in a cloned address space. Is it undone through the KEEPONFORK
advice.

This also implements MADV_NORMAL which is a no-op advice value used by
the ltp tests for MADV_WIPEONFORK and fixes a small bug in the order of
error checking on overflow. We're supposed to check for overflow on the
end address of madvise() before checking if the pointer value is out of
range for the address space.

Change-Id: I6e7ad2bb248d21348e03b8dde7cf5cea179b4a12
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/919753
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: James Robinson <jamesr@google.com>