[fidl][cpp] add fidl::IsWire<T>
This allows code to specialize based on whether a type is from the
natural or wire bindings. We need to do this because sysmem relies on
internal implementation details of the new C++ bindings to guess this,
and with the removal of the V1 wire protocol those are changing.

Change-Id: I6d2d734fecdd9d76b48e8ebdf5d8dfb779ee4e7c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/923134
Reviewed-by: Adam Barth <abarth@google.com>
API-Review: Adam Barth <abarth@google.com>
Fuchsia-Auto-Submit: Ian McKellar <ianloic@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>