[fshost] enforce tear_down in tests
Add a check to enforce calling tear_down on the test fixture before the
fixture is dropped, and add the call where it was missing.

This also revealed a situation where we were filing a crash report for a
corruption after a normal operation where we shred the keys, so some
refactoring is done so we don't file the crash report. Technically this
also fixes a bug where shredding the keys would break a device, if the
device was configured not to format the data volume on corruption.

Test: fx test fshost_integration_tests
Change-Id: I4ee47c8ca4760f1e0934bf6e79d8584f8de7d9c9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/922312
Reviewed-by: James Sullivan <jfsulliv@google.com>
Commit-Queue: Stephen Demos <sdemos@google.com>
Fuchsia-Auto-Submit: Stephen Demos <sdemos@google.com>