[benchmarks] Enable storage tests (and some others) on Nelson
Enable the tests that are currently run on Astro and Sherlock on
Nelson too.

For storage tests, it would be useful to see whether the changes in
"fxblob-pkgdir" performance metrics that we've seen on VIM3 and
Sherlock also occur on Nelson.  That would give us some clues about
why the changes occurred.

Bug: 133969
Change-Id: I2b01e6d202b7fb4e0a250b38c93b84735d3cc229
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/923912
Commit-Queue: Mark Seaborn <mseaborn@google.com>
Reviewed-by: Chris Drouillard <cdrllrd@google.com>