[superproject] Roll third_party/pigweed docs: Start the glossary
Creates a new glossary (//docs/glossary.rst) and starts it off with a
definition for "module". Updates other docs to link to the glossary
definition.

Over the long run we may want a more structured approach. Maybe we
store the glossary definitions in a different format (e.g. JSON) to
make it easier to auto-link to the glossary in the rest of the docs.
For now, we should just start with the simplest solution and only
add complexity when necessary.

Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/172952
Original-Revision: 7ea3c0814e1f56d9b1e3148d490ab1540070f1bb
GitOrigin-RevId: 63fe5b6f08617268be7a6e996e60a50be2632a8d
Roller-URL: https://ci.chromium.org/b/8768678551862459377
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I538ae3a2a8d94bb49079887a3d6174a4d9a1a51e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/924172