[bt][avrcp] Set browsed player after browse connection is set up.
After browse connection is set up, the controller sets the browsed
player. If a player is addressed, we set the browsed player to the
addressed player. Otherwise, we set it to a player with the highest
browse capability. The player will be set as browsable if it supports
of browsing.

Browse player is set in the peer not in the controller, so that
different controller instances (i.e. from ControllerService and from
BrowseControllerService) that point to the same peer can have the
already addressed browsed player information.

Change-Id: I74f61368590e95181417046e89dcda8ff4bc4275
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/830496
Commit-Queue: Dayeong Lee <dayeonglee@google.com>
Reviewed-by: Marie Janssen <jamuraa@google.com>
Reviewed-by: Ani Ramakrishnan <aniramakri@google.com>