[pb] Fix download when logic when remote is an absolute url
Many of the unit tests have TransferEntry.remote as an absolute url
(e.g. gcs://bucket/b), but this doesn't seem to actually work (in
production, TransferEntry.remote is always a relative path). It seems
like it could be useful to support an absolute path in the transfer
manifest, which would let us freely move the transfer.json while leaving
the artifacts in their original location. This change implements this.

Change-Id: I489403f27f53bdd6a3228687139644fd002f5d15
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/924292
Commit-Queue: Michael Brase <mbrase@google.com>
Reviewed-by: Jiaming Li <lijiaming@google.com>