[cpu-trace] Migrate To a Regular Component
cpu-trace doesn't need to be a driver anymore. It originally had to be in
order to get the root resource, but that is no longer the case.

This patch modifies the architecture of cpu-trace to be a
fuchsia.perfmon.cpu fidl server. Provided a DebugResource, it provides
an interface to the mtrace syscall. Then we make it a child component of
the frontends: cpuperf_provider and the cpuperf shell package.

Bug: b/294603854

Change-Id: I19d811f38d48878d2895439fa40f49c6490b87e9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/901300
Reviewed-by: Krithika Balan <krithikabalan@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Fuchsia-Auto-Submit: Gwen Mittertreiner <gmtr@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>