[wlan][hw-sim] Generate unique test-root for each test case
CTF will require each test suite subpackage to be uniquely named and as
a result hw-sim can no longer use the same test-root component manifest
that hardcoded the name of the test suite subpackage.

This changes hw-sim tests to generate a test-root manifest and
component for each test suite component based on the name of the target.

This also changes the generated names of the hw-sim test packages to
include "wlan-hw-sim" so that "fx test wlan-hw-sim" works again.

Test: fx test wlan-hw-sim
Change-Id: If0a5138af2fea25f932ac74f8847dcbb94ae08b7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/921055
Reviewed-by: Kendal Harland <kjharland@google.com>
Reviewed-by: Charles Celerier <chcl@google.com>
Commit-Queue: John Tan <johntan@google.com>
Reviewed-by: Sakthi Vignesh Radhakrishnan <rsakthi@google.com>