[fidl][compiler] Rework compiler diagnostics tests
This introduces ExpectFail and ExpectWarn methods on the TestLibrary
class and an ASSERT_COMPILER_DIAGNOSTICS helper macro that compiles the
library, checks that the diagnostics match the expectations and prints
any differences.

With these a test can assert more precisely which errors and warnings
should be generated by a compilation. Without these, when a test that
checks compiler errors or warnings fails it can be hard to understand
why.

This updates a couple of tests I had to address for unknown interactions
to ASSERT_COMPILER_DIAGNOSTICS but leaves most untouched. As we
experience test failures related to diagnostics we should update those
tests to use this approach.

Change-Id: I4db43b25f6e2fc8e4b5b70fad283b322e3a625d5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/924088
Reviewed-by: Mitchell Kember <mkember@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Ian McKellar <ianloic@google.com>