[amlogic-display][vout] Remove "VoutType::kUnknown".
kUnknown used to indicate that the Vout is not yet initialized
with any type of output. Now that the output type is always set
up at Vout class construction, it's impossible for a Vout to be
of "kUnknown" type, so we can safely delete it.

Test: no-op. Dead code deleted.
Change-Id: Id9e061d863cc508657aab93f4d27e73ae51ca65e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/924250
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Yilong Li <liyl@google.com>
Reviewed-by: Victor Costan <costan@google.com>