[ffx][build] Add runtime_files to tool_paths.json
This adds an entry to the tool_paths.json's tool items that points at a
list of runtime files that should be present when the tool is run. It's
used here to add the ffx subtool metadata file so that tools like
testsharder can find it without guessing at its location.

Bug: b/302726815
Change-Id: Ifef7ad6199a1e02379f93b9d71d997bff2637543
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/924752
Reviewed-by: Ina Huh <ihuh@google.com>
Commit-Queue: Megan Batty <mgnb@google.com>
Reviewed-by: David Turner <digit@google.com>